Нашел ошибку в своих вычислениях.
