+++ labs/lab3/example/.gitignore
@@ -1,0 +1,1 @@
+result/
