+++ labs/lab3/example/CMakeLists.txt
@@ -1,0 +1,3 @@
+cmake_minimum_required(VERSION 3.15)
+
+add_subdirectory(src)
